Separate model and validate checks
